feat:quran reading page

+++ app/Helpers/functions_helper.php
@@ -20,13 +20,14 @@ function setRandomColor() {
 }
 
 function toNumber($number){
-    return preg_replace("/[^0-9]/", '', $number);
+    return (int) preg_replace("/[^0-9]/", '', $number);
 }
 
 
 function toPhoneNumber($number){
     preg_match('/^0+/', $number, $matches);
     $number = toNumber($number);
+    // var_dump($matches, $number);
     $hp = toNumber($number);
     return ($matches[0] ?? '').$hp;
 }
